+++ g431-master/app/peripheriques/button.h
@@ -18,10 +18,10 @@ typedef enum
 void BUTTONS_init(void);
 void BUTTONS_update(button_e * button_left_event, button_e * button_right_event, button_e * button_up_event, button_e * button_down_event, button_e * button_center_event);
 
-uint8_t BUTTON_left_read(void);
-uint8_t BUTTON_right_read(void);
-uint8_t BUTTON_up_read(void);
-uint8_t BUTTON_down_read(void);
-uint8_t BUTTON_center_read(void);
+int BUTTON_left_read(void);
+int BUTTON_right_read(void);
+int BUTTON_up_read(void);
+int BUTTON_down_read(void);
+int BUTTON_center_read(void);
 
 #endif /* BUTTON_H_ */
